+++ resources/views/filament/resources/user-login-records/widgets/date-range-filter-widget.blade.php
@@ -1,7 +1,0 @@
-<x-filament-widgets::widget>
-    <x-filament::section>
-        <form wire:submit.prevent>
-            {{ $this->form }}
-        </form>
-    </x-filament::section>
-</x-filament-widgets::widget>